Remove Legacy Scrypt and replace with V2
